Extract MongoSSL changes to a dedicated module
