+++ plugin/trino-mongodb/src/test/java/io/trino/plugin/mongodb/TestMongoClientConfig.java
@@ -16,16 +16,11 @@ package io.trino.plugin.mongodb;
 import com.google.common.collect.ImmutableMap;
 import org.testng.annotations.Test;
 
-import javax.validation.constraints.AssertTrue;
-
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.Map;
 
 import static io.airlift.configuration.testing.ConfigAssertions.assertFullMapping;
 import static io.airlift.configuration.testing.ConfigAssertions.assertRecordedDefaults;
 import static io.airlift.configuration.testing.ConfigAssertions.recordDefaults;
-import static io.airlift.testing.ValidationAssertions.assertFailsValidation;
 
 public class TestMongoClientConfig
 {
@@ -42,10 +37,6 @@ public class TestMongoClientConfig
                 .setConnectionTimeout(10_000)
                 .setSocketTimeout(0)
                 .setTlsEnabled(false)
-                .setKeystorePath(null)
-                .setKeystorePassword(null)
-                .setTruststorePath(null)
-                .setTruststorePassword(null)
                 .setMaxConnectionIdleTime(0)
                 .setCursorBatchSize(0)
                 .setReadPreference(ReadPreferenceType.PRIMARY)
@@ -58,9 +49,6 @@ public class TestMongoClientConfig
     public void testExplicitPropertyMappings()
             throws Exception
     {
-        Path keystoreFile = Files.createTempFile(null, null);
-        Path truststoreFile = Files.createTempFile(null, null);
-
         Map<String, String> properties = ImmutableMap.<String, String>builder()
                 .put("mongodb.schema-collection", "_my_schema")
                 .put("mongodb.case-insensitive-name-matching", "true")
@@ -71,10 +59,6 @@ public class TestMongoClientConfig
                 .put("mongodb.connection-timeout", "9999")
                 .put("mongodb.socket-timeout", "1")
                 .put("mongodb.tls.enabled", "true")
-                .put("mongodb.tls.keystore-path", keystoreFile.toString())
-                .put("mongodb.tls.keystore-password", "keystore-password")
-                .put("mongodb.tls.truststore-path", truststoreFile.toString())
-                .put("mongodb.tls.truststore-password", "truststore-password")
                 .put("mongodb.max-connection-idle-time", "180000")
                 .put("mongodb.cursor-batch-size", "1")
                 .put("mongodb.read-preference", "NEAREST")
@@ -93,10 +77,6 @@ public class TestMongoClientConfig
                 .setConnectionTimeout(9_999)
                 .setSocketTimeout(1)
                 .setTlsEnabled(true)
-                .setKeystorePath(keystoreFile.toFile())
-                .setKeystorePassword("keystore-password")
-                .setTruststorePath(truststoreFile.toFile())
-                .setTruststorePassword("truststore-password")
                 .setMaxConnectionIdleTime(180_000)
                 .setCursorBatchSize(1)
                 .setReadPreference(ReadPreferenceType.NEAREST)
@@ -105,27 +85,5 @@ public class TestMongoClientConfig
                 .setImplicitRowFieldPrefix("_prefix");
 
         assertFullMapping(properties, expected);
-    }
-
-    @Test
-    public void testValidation()
-            throws Exception
-    {
-        Path keystoreFile = Files.createTempFile(null, null);
-        Path truststoreFile = Files.createTempFile(null, null);
-
-        assertFailsTlsValidation(new MongoClientConfig().setKeystorePath(keystoreFile.toFile()));
-        assertFailsTlsValidation(new MongoClientConfig().setKeystorePassword("keystore password"));
-        assertFailsTlsValidation(new MongoClientConfig().setTruststorePath(truststoreFile.toFile()));
-        assertFailsTlsValidation(new MongoClientConfig().setTruststorePassword("truststore password"));
-    }
-
-    private static void assertFailsTlsValidation(MongoClientConfig config)
-    {
-        assertFailsValidation(
-                config,
-                "validTlsConfig",
-                "'mongodb.tls.keystore-path', 'mongodb.tls.keystore-password', 'mongodb.tls.truststore-path' and 'mongodb.tls.truststore-password' must be empty when TLS is disabled",
-                AssertTrue.class);
     }
 }
